GS-63 remove util
