+++ server/src/middleware/data.ts
@@ -1,7 +1,34 @@
+import { game } from '../models/game.model';
 const AWS = require("aws-sdk");
 
 AWS.config.update({region: 'us-west-2'});
 let ddb = new AWS.DynamoDB();
+
+export function prepareGameAssets(playerCount: number): Promise<void> {
+  return new Promise((resolve) => {
+    const numPlayers: string = (playerCount <= 3) ? '3' : playerCount.toString();
+    const cardFilter = {
+      FilterExpression: "#np <= :np",
+      ExpressionAttributeNames: { "#np": "NUM_PLAYERS" },
+      ExpressionAttributeValues: { ":np": { N: numPlayers } }
+    };
+    console.time("Load Data")
+    const boards = Promise.resolve(loadTable('BOARDS', 'BOARD_ID'));
+    const cards = Promise.resolve(loadTable('CARDS', 'CARD_ID', cardFilter));
+    console.timeEnd("Load Data")
+    Promise.all([boards, cards]).then(() => {
+      resolve();
+    });
+  });
+}
+
+
+function loadTable(tableName: string, id: string, params: {} = {}): Promise<void> {
+    const tableData = Promise.resolve(tableScan(tableName, id, params));
+    return Promise.resolve(tableData).then((response) => {
+      game[tableName.toLowerCase()] = response;
+    });
+}
 
 // Print credentials to confirm they are correctly configured
 // The location of the shared credentials file 
@@ -22,9 +49,8 @@ function tableScan(tableName: string, tableID: string, addParams: {}): Promise<a
             TableName: tableName,
         }
         ddb.scan(params, function(err: Error, data: any) {
-            if (err) {
-                console.log("Error", err);
-            } else {
+            if (err) console.log("Error", err)
+            else {
                 console.log("Success: Retrieved " + data.Count + " entries from " + tableName);
                 let convertedData: any = {};
                 data.Items.forEach((record: object) => {
@@ -36,5 +62,3 @@ function tableScan(tableName: string, tableID: string, addParams: {}): Promise<a
         });
     })
 }
-
-exports.tableScan = tableScan;
