+++ server/src/middleware/gameFeed.ts
@@ -1,5 +1,4 @@
-import { pushUpdateToPlayers } from "./util";
-import { game, serverData } from '../models/game.model';
+import { game, serverData, pushUpdateToPlayers } from '../models/game.model';
 import { feedItem } from "../models/playerData.model";
 
 // --------------------
