+++ server/src/models/game.model.ts
@@ -33,3 +33,14 @@ export let serverData: any = {
   clients: [],
   gameCountdown: undefined,
 }
+
+let sseId: number = 2;
+
+export function pushUpdateToPlayers(data: string, event: string = 'message', clients: any) {
+  clients.forEach((client: any) => {
+    client.res.write(`id: ${sseId++}\n`);
+    client.res.write(`event: ${event}\n`);
+    client.res.write(`data: ${data}\n\n`);
+    client.res.flush();
+  });
+}
